+++ server/src/main/java/imag/quizz/server/network/PlayerConnectionManager.java
@@ -4,7 +4,7 @@ import imag.quizz.server.PlayerController;
 
 public class PlayerConnectionManager extends ConnectionManager {
 
-    public PlayerConnectionManager(final PlayerController controller, final int localPlayerPort, final int ownId) {
+    public PlayerConnectionManager(final PlayerController controller, final int localPlayerPort, final long ownId) {
         super(controller, true, localPlayerPort, ownId);
     }
 }
